+++ internal/params/parameters.go
@@ -281,6 +281,27 @@ func (p *Parameter) UUID(v [16]byte) Builder {
 	return p.parent
 }
 
+func (p *Parameter) TzDate(v time.Time) Builder {
+	p.value = value.TzDateValueFromTime(v)
+	p.parent.params = append(p.parent.params, p)
+
+	return p.parent
+}
+
+func (p *Parameter) TzTimestamp(v time.Time) Builder {
+	p.value = value.TzTimestampValueFromTime(v)
+	p.parent.params = append(p.parent.params, p)
+
+	return p.parent
+}
+
+func (p *Parameter) TzDatetime(v time.Time) Builder {
+	p.value = value.TzDatetimeValueFromTime(v)
+	p.parent.params = append(p.parent.params, p)
+
+	return p.parent
+}
+
 func Declare(p *Parameter) string {
 	return fmt.Sprintf(
 		"DECLARE %s AS %s",